Refactor imports, adjust component positions, update API endpoint
