+++ app/dashboard/[workspaceId]/application/[appId]/edit/page.tsx
@@ -1,11 +1,5 @@
-"use client";
-import React, { useEffect, useState } from "react";
-import { CSS } from "@dnd-kit/utilities";
-import { useDraggable, useDroppable } from "@dnd-kit/core";
-
-type Props = {};
-
-// create array of nested id's of components1
+import React from "react";
+import Editor from "../../_components/Editor";
 
 interface ComponentData {
   id: number;
@@ -14,30 +8,32 @@ interface ComponentData {
   content: string;
   // Add more configurable properties as needed
 }
+type Props = {};
+
 const test: ComponentData[] = [
   {
     id: 1,
     content: "Component 1",
-    x: 174,
+    x: 1274,
     y: 415,
   },
   {
     id: 2,
     content: "Component 2",
     x: 574,
-    y: 475,
+    y: 1175,
   },
   {
     id: 3,
     content: "Component 3",
     x: 18,
-    y: 72,
+    y: 822,
   },
   {
     id: 4,
     content: "Component 4",
     x: 877,
-    y: 459,
+    y: 14259,
   },
   {
     id: 5,
@@ -47,63 +43,12 @@ const test: ComponentData[] = [
   },
 ];
 
-const Draggable = ({ id, content, x, y }: ComponentData) => {
-  const { attributes, listeners, setNodeRef, transform, isDragging } =
-    useDraggable({
-      id: id,
-    });
-
-  const style = {
-    transform: CSS.Transform.toString(transform),
-    position: "absolute" as const,
-    left: x,
-    top: y,
-    opacity: isDragging ? 0.5 : 1,
-    background: isDragging ? "lightgreen" : undefined,
-  };
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={style}
-      {...attributes}
-      {...listeners}
-      className="bg-white rounded shadow-md touch-none"
-    >
-      {/* Your content here */}
-      {content}
-    </div>
-  );
-};
-
 const page = (props: Props) => {
   return (
-    <div className="w-full h-full">
-      <div>
-        {test.map((item) => (
-          <Draggable key={item.id} {...item}></Draggable>
-        ))}
-      </div>
+    <div>
+      <Editor />
     </div>
   );
 };
 
 export default page;
-
-// const Droppable = ({ id, children }: any) => {
-//   const { isOver, setNodeRef } = useDroppable({
-//     id: "droppable",
-//   });
-//   const style = {
-//     color: isOver ? "green" : undefined,
-//   };
-//   return (
-//     <div
-//       ref={setNodeRef}
-//       style={style}
-//       className="bg-white/20 w-full h-full p-2 rounded-lg text-center"
-//     >
-//       {children}
-//     </div>
-//   );
-// };
